This is not working. Even on just the internal parts, the SM update is too unstable. Good thing thatI did not start out by writing those optimized kernels, eh?
I have no choice, but to move on. I'll try to make it work with explicit inverses. I'll have to make use of the Cholesky decomposition and the triangular solve in order to implement it.

After I've made sure that PRONG works, only then I will move on to looking into ways of regularizing the SM updates.

I'll have a break here. I've been at it for a while.