I had a really good idea just now so I had to spend some time thinking about it.
I understand how to get rid of the need for the inverse entirely. Yeah, basically there is no way to make the inverse update stable. I can just abandon all hope with regard to that.

But there might be a way to get close to the K-FAC update without needing to do an inverse.

I need to go through the starter lecture for differential geometry again. I forgot how to do second derivatives and need it for this.

It is really easy to describe what the update should be. The K-FAC update is `M = U^-1 x g R^-1`.

This can be rewritten as: `U M R = x g `.

The gradient of `M` on the first step would be `U x g R` assuming `M` is set to zero. This is what I could reduce the backward step into. But this is just the first derivative. Let me try and figure out what the second step would be. That would reduce the error.