I am tired of Dark Souls at this point. That game is kind of super easy now, in fact I am really getting bored with it.
Let me write a bunch of things.

Right now I am really bothered how the last 3 months went by. I am really making pitiful progress. I could have implemented K-FAC in a week back in April had I wanted to. Instead I am working my way towards dumb ideas like the one I had today. Sure you could set the M to zero and do a one step update, but isn't that really stupid?

Well, I'll give it a try since I do not have anything better in mind, but I think that past this point I should just focus and do PRONG with k step updates. Apart from Cholesky, I'll have to put in the LU decomposition for Zap Q as I'll need it for the inverse, but I think that unlike QR the LU could be adapted to row major layouts.

Of course I could use more than one step update, but in that case from both precision and a performance standpoint I'd be much better off going with a direct decomposition.

I wish linear solvers could do the PRONG case in an iterative fashion, but I do not think numerical stability is that important here. And if I end up repeatedly using the inverses, then it would be better to actually make that step.

Ah, forget it. I am throwing in the towel and simply going with PRONG as flow dictates. I am absolutely certain that there will be a iterative way of doing it in the future, but I might need to leave this to the AIs.

So tomorrow it is PRONG day. It is going to go in and then it will be followed closely by Zap Q. I'll also implement all the symmetric operations as needed. That is what Spiral is for and what can be expected of me. Not huge mathematical insights.

At that point I will be sitting good, and in a comfortable place. I'll let the whole Rudder story play out as it needs. I need to dive into poker. There is a lot of work to do and I need to go through it. Let me aim to finish all of this in the next month or two and then I will be able to take a break while the agents are training. I am going to need to consider how exactly I want to interface with the targets.

This problem of interfacing is something that would be really fitting to somebody of my skills and I could definitely make contributions there. Trying to figure out improvements for PRONG is just torture. I think I am at the point where I can acknowledge my spirit, but my math talent has never leveled up despite my effort.

Let me just say one thing. It must have been like 1.5 months ago now, but I was right to doubt PPO. It is not the kind of the thing that would ever be a substitute for natural gradient methods. After all my learning that piece of doubt towards it is definitely vindicated and is a good indication not to buy into a story as presented blindly.

Damn I am tired. It is 10:40pm at the time of writing this.

I'll be heading to bed soon and during the night I'll try to catalyze a change in my mindset. Those linear solvers and those decompositions are there for a reason. While push into math gave me the great insight that PRONG implies K-FAC, anymore than this and I will be just frittering away time and not facing my direct goal. It would just be cowardice. I definitely did enough exploration.

Tomorrow I will lift my seal and go 100% into programming rather having my effort split between 20% programming 80% research.