Since Sherman Morrison update is a failure for updating the inverse of the covariance matrix, I'll leave the work that I did on a separate branch just in case I need it later.
https://www-users.cs.umn.edu/~saad/IterMethBook_2ndEd.pdf
(book) Iterative Methods For Sparse Linear Systems

My next plan and my last hope is to try out the weighted Jacobi iteration. If that does not work I will do it like the standard PRONG does and use a direct decomposition to get the inverse.

I'll see what the book I linked to has to say about it and begin work on the tests.