I am done with iterative updates.
The weighted Jacobi iteration does not work. I knew that all these methods had flaws, but I was hoping they would work well in practice. Unfortunately they do not.

Let me make a tally.

1) RNG L2 update - does not work.
2) Inverse Cholesky update - does not work.
3) Sherman Morrison update - does not work.
4) Weighted Jacobi iteration - does not work.

What a waste of time the last month and a half was. I should have saved myself 5h today and simply did what I said I would in last night's commit which was to bring in the direct inverse.

Ultimately, these updates are supposed to be the most stable part of the whole architecture, but they just keep dying even on toy examples.

This is no way to do ML.

I'll put the code for this on a separate branch and then start from a clean slate.