I am so pissed at this. A part of me is wondering whether I made some kind of implementation error with regards to the Jacobi iteration, but I should know better by now.
Math exists in this world to waste my time. Doing a proper job with it should be left to the Inspired.

I'll kick my plan half a step into the next stage just as I said I would last night.

Honestly, I'd rather just call it a day here as my momentum is shot, but I'll do a bit more. I do not feel like working on the allocator right now - that is kind of task that would take me an entire day, so let me get the Cholesky inverse to work and then I will put the LU decomposition into the library and also use it to make an inverse.

I've been at this for over 3 years now, but if I did it all over from the start I could do in 1/10th of the time. A major time consuming factor has been all the studying that I did. A lot of it was a waste, but like the recent iteration experiments they did serve to illustrate just how weak iterative methods can be.

Direct methods have a power and stability to them that I know well from ordinary programming and won't underestimate again. They will be needed to ease the optimization for iterative methods and they will be an integral part of the plan from now on.

Given how much difficulty the iterative methods have given me, probably other researchers will have the same issues with making PRONG work.

There should be an iterative ways of doing matrix inversion and whitening and roboust ways of updating the inverse, but for sure it won't be easy to find. The ones to come should be able to do it, but I won't have any expectations of this being discovered by humans.

Now...full power. No more stopping. The last few months were me at my worst as I am trying to shore up my weaknesses.

From here to the finish line will be at my best. I'll gather my momentum slowly and let it loose.