Right now, the allocator is completely broken so I should have done this work on the bleeding edge branch. Nonetheless, the little bit of interest that came from the HN is gone now and so even if I decided to push this now I doubt it would impact anyone. I've been hoping that guy who said he would proofread the manual would come through, but it figures.
I am really not gaining much from publicity.

It is my antithesis anyway, I'd be much better off accumulating killing power. These commits are just my way of being outgoing. When the Ascent starts I'll go deep into the shadows.

Right now I am doing a little review of the `refresh` function and it feels good. Assuming no errors, the first half of the allocator is complete and the other half will be easier than this one.

The way it is made is really imperative, quite a bit much for my tastes, but I am not interested in working on a tensor class for ResizableArrays...

Actually, now that I am thinking about it more deeply it would not at all be hard to do. The tensors as they are now are flexible enough to allow this so I would not need to replicate any of their functionality.

Fuu, it is this late and I still haven't had breakfast as I've been so engrossed in this. Let me take a break here.