I got up an hour ago and have been chilling. I've been thinking about the RNG approximation, but looking at it again, it is as just as difficult to implement as I remember it.
I think the EM way of approximating the inverse might have better optimization properties, but still, it would be quite expensive.

Matrix multiplies are not cheap in the L2 cost, and the log det version is just a pain in the ass.

I am really curious as to how expensive the Cholesky decomposition will turn out to be. Going by the O notation,  it should be something like 2x the times of a standard multiplication. So the inverse should be something like 3x.

But I am really not sure how expensive it will be in terms parallelism. It might be the deciding thing that makes it a lot more expensive than expected.

Well, it should not be too bad if I do it every 10 steps or so.

One thing I am sure of is that it will be a lot less expensive than the eigendecompositions the PRONG papers are using. It is definitely better idea than that.

So I guess that would be another gain on my side. Along with the K-FAC connection, I think I should just be satisfied with that. Replacing the eigendecomposition is not a trivial thing.

Neither is the insight how to do it for multiple layers.

Ah, there is one thing I am going to have to do. I am going to need some kind of neuron pruning. Since the Cholesky decomposition will give me the eigenvalues, I really should block the low ones, not just to improve generalization but because the units with very low eigenvalues will break the net if they ever get used since their inverses will be very high.

Enough of this.

Let me just put in a few words on the allocator. It is not completely tested, but the same would apply to the previous version. It is actually difficult for me to construct tests that would stress it in a realistic environment so some code paths will just have to be left alone for a while, but I have a good feeling about this. The way it worked the first time once it compiled is definitely a very good sign. I do not think I missed anything, but I will have to keep in mind that it might not be fully stable yet.

With that out of the way, let me get to actual work.

I do not need to put in the solver function from cuSolver. I know that the inverse will never fail if the Cholesky itself does not.

So that leaves PRONG. I will finish it today. There is nothing at all stopping me anymore, so how about I finally get to it. It should has been a winding road to get to it.

Let me check out how well it does in practice.