Here is the power function.
I think I basically have it with this. All that remains is to get it to compile.

I think for this I am going to cut the size of the training set to 1/10th because it might take long otherwise.

Let me get on with it. I'll make the updates stepwise later.