Let me set the minibatch size to one. Maybe rescaling the gradients rate on the backwards step is what is messing with it.
I did not do this well. It seems there is a serious problem with dividing directly on the top level.

```
Unhandled Exception: System.Exception: The leading minor of order 214 is not positive definite.
```

Oh, maybe I need to increase the default epsilon.

```
Unhandled Exception: System.Exception: The leading minor of order 214 is not positive definite.
```

What the fuck. This is with even an epsilon of 0.1. There is something messed up going on here.

...Even with epsilon of 1.

...I need to print this out.