"8:45am. I've tested the k steps. The conclusion is that I can freely make it 256 without any loss in performance. I could go even more, but I do not see much point.
I pretty much wasted the last one and a half month. Let this be a lesson.

Let me speed up the multiply of the precision matrix by the use of a symmetric multiply.

8:50am. https://docs.nvidia.com/cuda/cublas/index.html#cublas-lt-t-gt-syrk

I need to put these optimizations in. I should also put the regular matrix vector multiplies as well.

I guess doing this stuff will be my mission for the rest of the day. I want to speed up the updates and there is still room to do so.

I need to make `portf` and `trinv` in-place.

Right now they are just wasting memory.

9:05am. Well, let me start off by putting in `syrk`."