I am nearly done with the PRONG arc. Now I am running it on the full dataset in online mode. For 3 hidden layers, it seems PRONG likes a lot lower global learning rates than SGD does. I guess I do not need to worry about superconvergence with it.
But damn online learning is slow. This is what I can expect when doing RL. I wish I could avoid this somehow, but PG are on policy. The GPU really is not the best choice for RL, but it will have to make do.

I think I should try an iteration with a low learning rate and a few with higher ones.

I'll also compare it with standard SGD.

Looking at the training vs test error, it seems that 3 iterations is enough to make it fully converge. I do not think this will be the case for SGD.

Yeah, forget fiddling with raising learning rates. That will just make it converge that much quicker on the test set.