``` Training: 1.16576913669581 Testing: 0.425718714889043(8884/9984) Training: 0.285756380478784 Testing: 0.319342847435902(9050/9984) Training: 0.148596439994816 Testing: 0.295945314750171(9084/9984) Training: 0.0730274195579082 Testing: 0.288669545788509(9114/9984) Training: 0.0337782487770184 Testing: 0.286512328926713(9167/9984) ```
For 3 hidden layers, it seems PRONG likes a lot lower global learning rates than SGD does. I guess I do not need to worry about superconvergence with it.

Now, let me use the whole dataset.

```
Training: 0.290151304830563
Testing: 0.112736815479226(9635/9984)
Training: 0.0701474280644621
Testing: 0.0962042397243353(9671/9984)
Training: 0.0283506842860905
Testing: 0.095039942302556(9699/9984)
Training: 0.0143732029160202
Testing: 0.0918081507466596(9732/9984)
Training: 0.00776721336608317
Testing: 0.0908223605719962(9741/9984)
```

It converges to a very low training error rather quickly.

Let me try SGD.

```
Training: 0.668517613038808
Testing: 0.813119509329016(8812/9984)
Training: 0.683598761678608
Testing: 0.610368761974267(9034/9984)
Training: NaN
Training diverged. Aborting...
```

Let me try lowering the learning rate for SGD by 3x.

But at first glance it seems the rumors were true. PRONG really is a lot better in the online regime than SGD.

```
Training: 0.233434581714346
Testing: 0.191192456115133(9395/9984)
Training: 0.110080066482164
Testing: 0.122853330826052(9614/9984)
Training: 0.0720736895353168
Testing: 0.106926092849626(9670/9984)
Training: 0.0487168198937373
Testing: 0.103757469533519(9704/9984)
Training: 0.0376629822915681
Testing: 0.101032782358323(9731/9984)
```

Or maybe not. PRONG seems to be just a bit better.

Unlike with batch learning, it seems that when doing online learning there is not much difference in speed. Let me time it.

```
Starting timing for: iteration 0
The time was 00:01:17.8073313 for: iteration 0
Training: 0.290151304830563
Testing: 0.112736815479226(9635/9984)
Starting timing for: iteration 1
The time was 00:01:17.6729953 for: iteration 1
Training: 0.0701474280644621
Testing: 0.0962042397243353(9671/9984)
```

This is PRONG.

```
Starting timing for: iteration 0
The time was 00:00:33.7830372 for: iteration 0
Training: 0.686073014714406
Testing: 0.387254213866515(8948/9984)
Starting timing for: iteration 1
The time was 00:00:33.7991602 for: iteration 1
Training: 0.360779539257332
Testing: 0.316935149713969(9099/9984)
```

This is SGD. I forgot to tune the learning rate.

So the difference is only on the order of 2.5x.

But regardless this is really horrible compared to minibatch learning in terms of runtime.

Since I will be doing a lot of online learning, what I will do is bring back the matrix vector products. I'll have syrk, symm, trmm and gemm route into them when one of the arguments is a vector. I know there is a 3x difference in speed between level 2 and level 3 functions. I need to take advantage of it. Right now the GPU is just cooling.

Let me see if I can lower the default regularization a bit and then I will be done for now.

Putting in the optimized paths will be my mission for the rest of the day.

After that is done, I am done with PRONG. It will be like any other component that I can just plug in. I'll resume work on RL and aim to create an actor-critic architecture. Once that is done, I'll implement Zap Q.

I do not feel like bothering with multi layer updates since it overfits enough as it is.

August definitely feels like it will be a productive month for me. In the last two days it feels like I made more progress than in the rest of the days in the last two months.