I am having difficulty putting my thoughts into words, but I will try.
What I want to do here is do some radical cutting on the ML library. There is really a lot of stuff in it that while interesting to implement at first, is now a dead weight for me.

The point of contention would start at the action layers, but now that I am going over it, I see that the rot is everywhere.

There is a very good reason why I had to dump all my old for q learning and whatever is in those text files in the Learning project. This is a very real indication that there is something wrong with the high level structure of the library.

The third point is the sampler. Somehow I need to append a layer at every point in order to pass the inputs into it. This would not be bad except for the tiny fact that it creates GIDs which are the only impure aspect of Spiral's otherwise pristine type system.

...