A library of well kept primitives is always an asset, but high level structure gives way to disfunction.
Like if I ever want to add dense residual connections that would be a major pain in the ass with the current structure.

Edit (2h later): I am still thinking about it. It has been a while since I really last did this. This is more my cup of tea than trying to peer inside the soul of linear algebra operations.

There are definitely a bunch of things I do not like in the ML library.

The way I approached it made sense originally and with regard to what I wanted to do back then, but what I wanted to do back then was not a really good thing to want.

Some of the things like the state passing thing for the char RNN objectively have a large negative value to the library. I'd be better off simply increasing the sequence length.

The region passing technique I used there to pass state around for recurrent nets should have never seen the light of day.

The kind of thing is simply impossible to reason about. I hated it even back then and it made me sorely wish for GC for GPU memory.

It is going out and I'll be glad to get rid of it.

Of the things that I want to get rid of, layer norm would be one. It was one of the major motivators for me creating Spiral, but I do not need it anymore now that I have PRONG. It is just taking up space.

I am going to be taking global IDs out of the language. If I am unlucky somebody is going to end up using that stuff and making a mess. All compile time stuff in Spiral apart from function memoization should be pure.

Let's see...even ignoring the above dislikes the library does have a huge massive lack that would in fact necesitate a complete overhaul eventually. The lack of model checkpointing capability.

I did not need it back then since I was fluking around on Mnist and tiny-Shakespeare, but I think it is time to gradually move to bigger leagues.

This is the ideal time to bring in these changes.

Things like the wavefront iteration I simply do not need and I've been hamstrung by the lack of flexibility in areas I did not even expect.

What I will do is dump the library to the side and cherry pick the things off its corpse much like I've been doing from the F# ML library from 2 years ago. I still use it as a reference from time to time.

I will rebuild from the ground up, aiming to chisel it down to the bare bones.

...Damn it, even now that state passing trick haunts me. There is really no way around it but full GC for GPU memory. I need to force it deep down and focus on the task at hand. I simply won't ever need it again in the context of NNs.

How annoying.

Is this it? Do I have it all on hand now? Am I ready to start?

Well since it is nearly 5pm obviously I won't be doing any programming today, but conceptually what I want to do seems straightforward.

I spent a lot of time crafting these high level structures, but they all reduce to a bunch of maps and reduction operations. A large part of my path of getting better at programming is getting better at reducing it to those functional operations along with knowing how to optimize them.

I think I have enough threads collected to start weaving.

Tomorrow I will begin work on the new library with the aim of redoing the  Mnist, char-RNN and sampling char-RNN tests. This should not take me more than a few days and at the end of it I will have a significantly simplified library that allows checkpointing.

After that I will bring the PG player and add a critic to it. Then I will implement Zap. Honestly, Zap is like half an hour of work to implement. It is pretty much trivial compared to the redesign. I expect the testing in comparison to take far longer.

After I have that, I will have to start getting fancy. I am not sure what I am going to do, but I am going to have to do something in order to make RL go faster. Online learning will not be cutting it, but at the same time I am not sure that batch learning is the answer here. I spent so much time on getting a grip on the GPU, but it is not the right device for the job. It is fine for supervised learning, but not so much for RL.

I think I might need to crib some tricks from the OpenAI guys working on ACKTR. Everything I can think of right now has such glaring flaws. I am not comfortable with any of the choices I can think of.

...Hmmm, an idea is coming to me. I am going to have to make special considerations for Prong so that the covariance matrices are not getting updated while the inversion is taking place, but apart from that I think those would be fine if I updated them either asynchronously or semi-asynchronously using atomics.

I'll have to take special note of which data structures are thread safe on the .NET side - I think that `ResizeArray` in particular is, but I am not sure about other things. I am going to have to study async programming a little.

Async will raise issues with regions too. I am considering paring them down so that one pointer can belong to one region only. That would make it much easier to deal with them.

I suppose I could do synchronous updates, but a major contributor to the bottom line will be updating as the agent plays. This is one thing I want to do at all cost. It is why I've gone through all these pains to study natural gradient methods and look for more efficient ways of doing updates of the inverse covariance matrix.

Online training is perfectly doable assuming I can cut variance down as much as possible and higher order optimizers are an essential part of that. With SGD though, the performance would fall of the cliff in the RL domain.

To that I will also add the differentiable plasticity type architectures as my RNN of choice.

Higher order optimizers, checkpointing, self play, asynchronous training, memory-based architectures - once that is set I will be able to seriously consider how to interface the agents I've trained with the world at large.

Now that will be a profitable endeavor.

Right now I still have the GTX 970 from 3.5 years ago as my main device. I have no intention whatsoever of letting the upgrade cycle grow to 8 years which was the case last time.

I think if I apply myself I could hit the next stage in less than two months. Wouldn't it great to have something of substance to report by the time the new year's review rolls around?

Let me end this commit here. Today was mostly a brainstorming session. It is an important part of programming to spend some time agonizing over the decisions before committing to them.