Let me take a break here.
The two most important functions are definitely the init and run. The rest are just complications I introduced misguidedly.

With this already have enough to start putting in layers. Also it slipped my mind, but I do not actually need the descriptors to do checkpointing. I could have done it with the previous design as well. All I really need is to iterate over the layers in turn.

Piece of cake. Writing serialization functions won't be a problem at all. Spiral is the best possible language for that sort of thing.

After I get back, I'll begin work on the layers and do the Mnist example. After that I'll jump straight into policy gradients. Nothing will stop me, nothing.