Let me just make a little commit to say just how great Dark Souls is. I am already on my third playthrough. I heard good things about it years ago and only got it because I saw it as I was browsing Skidrow.
Everytime I play these games I get closer to my old self and my true essence.

Literally the only flaw the games have is that it is impossible to draw power and bring it into the real world from them.

The solution to that is complete integration. The agents will have the capacity to draw power from games, and I will have the capacity to draw it from my agents.

That is the way it should be. That is what life in the 21st century is really all about. It is the long road leading to the Singularity.

I cannot accept what normal people see as obvious, that the games are worthless and a waste of time.

Just consider what they are - without a doubt, they are the pinnacle of art. They are little universes in the making. It cannot be allowed that they be subordinate to the social.

Damn I am tired. Tomorrow without a doubt I will get the PG agent online again and then implement the actor critic.

And the day after that I should be done with Zap Q. I fully expect this to work well. I will really be shocked if it does not.

The last three months were study and research. The next ones will be all about engineering.

I talked about the Singularity, but slowly I am going to start getting my own plan into gear.

I spent 2015 on practicing my programming skills, 2016 on making a ML library and 2017 on making Spiral. And the first half of 2018 was me finishing Spiral and forming my views on reinforcement learning.

Right now, I am very much complete and not held back by tools, my own understanding or lack of programming skill.

Right now I am truly at my best.

I'll do some really wonderful programming in the coming months. I guess it was destined that I would return to programming. It was always the thing closest to what my actual talent is connected to in the real world.

Had somebody else written Simulacrum in 2004 and shown it to me, I'd never have turned my back on it. I'd recognize the true message of the story right away.

If I had Spiral in the 00s, what would I have accomplished by now?