I wanted to start coding, but I had the same problem as two days ago - the need to think.
At first I was completely lost as the code is so convoluted, but after that initial perplexity I started to put things together.

The most important thing in understanding code is not to understand what it is. It is to understand what it should be.

Right now I need to put some thought into this. I might need to do some thinking here.