How about I call it a day here?
I'd really hoped to continue the momentum from yesterday, but today the imperative drive was to think. Much like two days ago I just ended up sitting in this chair for hours on end thinking.

I am whether I will resume programming tomorrow or end up just brainstorming again. Well, whatever the case I'll just let it play out in the only way it can. It is not like I have anyone breathing down my neck to do it one way or another. It will get finished eventually.

In my brainstorming, I realized that a lot of the difficulty facing me here is dealing with union types. So far, I've been able to get around the need to deal with them directly by CPSing and inlining, but all those efforts are strewn across the libraries and I do not have a module from which to draw from.

The (sparse) one hot serializer is just one aspect of having to deal with union types.

I've been thinking about it for a while, but today I finally figured out how to also do a dense serializer and how to represent integers in a range. Rather than figure out, it would be more precise to say that I've decided on a convention for that.

By its very nature, the game is forcing me to adopt a more imperative style and I have to deal with that.

The things that are blocking me here is not the net - it is having to manage it in a imperative context. Implementing PG and Zap is trivial.

Today I've been in the zone, but it was more like a design zone than a coding zone.

When I am done with the thinking, I think rather than work on PG, I will instead do a lot of cleaning up and work on type management, serialization and typeful loop unrolling. I want to do it all under one umbrella.

The unfinished players I've implemented here are sort of a sneak peek in how this will turn out.

The way they've been implemented is too much welded to the poker game. Implementing them in a generic fashion will lower the mental burden of managing them significantly and that is what I am driving for.

Once I get all of the friction out of the way the RL agents for poker, and not just it, will fall into place on their own.