"Let me extend the language with module_length.
```
[{index = 6; length = 8}, {index = 0; length = 0}, {index = 5; length = 0}]
```

Not quite what I had in mind. Let me fix the error.

11:55am. Well, I missed a apostrophe in `case_foldl_map`. It works now.

Let me take a break here. I'll work on the `from_sparse` after this. I really need to study the decode function for a while in order to figure out how it works again. It completely faded from my memory. The work on the `Union` module will not go by quickly. It will take some serious programming, no doubt about that.

But I have no doubt that this is what I should be doing."