It has been a while since I last worked on the language.
I got the idea to work on the `case_foldl` more when I stepped away from the screen to think a little. I was intending to figure out the semantics of dense serialization when this cropped up. It turns out I had done it incorrectly the first time around.

Thinking is never a bad thing it seems. Once I let my mind wander who knows what it will unearth.

To be honest, half an hour ago I started to feel restless as I was working on the dense function and that irritation to close for the day no matter how slight is always to be obeyed. Once you just want to rush through things you will 100% get killed in programming much like in Dark Souls. That having said, it does not happen at all often to me.

I am really really good at avoiding this trap at my age. I am not sure whether it was like that when I was a kid, but I've matured in some respects.

I'll dedicate time to properly doing the dense function tomorrow. Now it is time to get lunch and rest for the remainder for the day.