This is more difficult that I thought it would be. I needed to step away again to think a little.
Even though it is just a simple function for a bit there it went over the threshold of complexity that I am able to comprehend. I know how to simplify it though. I am going to do the same thing for `from_sparse` and get rid of that laziness in it.