Yeah, this is it. These rewrites will make it much easier to implement `from_dense`.
But not right now. I need a break here. As is usual at this time of the year, there is a heat wave and I need a refresher at this point. Maybe I'll finish this tomorrow if I do not manage it today. Probably I'll implement it today when I get back and leave testing for tomorrow.