https://www.reddit.com/r/ProgrammingLanguages/comments/93jfyz/august_18_monthly_what_are_you_working_on_how_is/e3gp44k/?st=jkcc8hwy&sh=fa3ed1de
Well here is the post on the PL sub. These commits and Reddit posts are a chronicle of my journey before the Singularity.

Now, to get to that full actualization I require more hard training. I think it is time that I did some work finally. Let me make the form function for from_dense.