Ahhh, I fucked up. I need to do the entire module all over again. Especially the `to` functions.
I know how I am going to do it. What I did in the current design is just stupid. Instead of having these special rules for integers I should have just introduced a bound integer type. With that I would not need to pass in templates into the `to` functions.