Let me call it a day here. I am done with the redesign. Tomorrow I will go through the tests again.
I am honestly quite please with how today went. These functions are by no means trivial, and the effort on them will be well spent.

After I am done with these I will do a bunch of work to bring in function unrolling with both inlining and union types into the Union module. That will make it really easy to deal with all the imperative coding that is facing me in the RL segment.

Well, that is enough of that. Let me have some fun.