"12:55pm.
```
    inl Rep = type {pot=int64; chips=int64; hand=Option.none Hand}
    inl State = Tuple.repeat num_players Rep
```

I need to change these into a bounded integer. Let me work on the game for a bit. I will tighten all this up.

12:55pm.

```
    inl Rep = type {
        pot=Union.int {from=0; near_to=max_stack_size*2} int64
        chips=Union.int {from=0; near_to=max_stack_size} int64
        hand=Option.none Hand
        }
```

This seems fine.

Agh, why did I do that `type_is`? I could have just used a `type` in the union tests.

Well, it is fine. The mistake was that it never occurred to me to do it without having to extend the language.

1pm. No, actually, no. I need to remove this from the language. Crap. I really was not thinking properly."