This is so damn annoying.
I got the tabular player to work, but it is only winning 66% against the rules based player instead of 72%.

There is something wrong with it.

But I am not sure I feel like dealing with this right now. In depth debugging is a very effort intensive activity and my stress level is high right now.

The biggest problem is not knowing what to doubt. I guess the most natural thing to doubt at this point would be the serialization functions. I do not want to doubt those, but I am going to have to.

Before I can move on, I need to be absolutely sure that they are spotless and that means no performance discrepancies between this and the previous version.

I am thinking of calling it a day here as much focus is busted, but let me take a break for a bit and I'll see if I can continue later.