Eureka! That was it after all. Immediately the win rate goes to 72.5%. There is no hideous bug lurking in the `Union` functions.
Good work me. I'll call it a day with this. It feels like it has been a while since I so heavily coded. Today's coding was definitely heavy somehow, I do not think it is usually like this.

With this the stage is set for policy gradient players to be introduced once more.

Making that a reality will be my mission tomorrow. Then I will move to creating actor-critics as per plan. I am not sure how I will fit Rudder into this whole framework, but it should come to me by the end of the year. I do really think that Rudder is one the most important RL papers of the deep learning era, but I do not yet know how to quite make it fit.

I need some actual insight on how AC works in practice and on real games, not one card poker which is a toy game.