"9:45am. Yeah, it does seem like SARSA works much better than Monte Carlo.
```
Starting timing for: iteration 0
Winrate is 34828 and 65172 out of 100000.
The time was 00:00:00.3554315 for: iteration 0
Starting timing for: iteration 1
Winrate is 50582 and 49418 out of 100000.
The time was 00:00:00.4577689 for: iteration 1
Starting timing for: iteration 2
Winrate is 56616 and 43384 out of 100000.
The time was 00:00:00.5179018 for: iteration 2
Starting timing for: iteration 3
Winrate is 64585 and 35415 out of 100000.
The time was 00:00:00.6239599 for: iteration 3
Starting timing for: iteration 4
Winrate is 66924 and 33076 out of 100000.
The time was 00:00:00.6579779 for: iteration 4
Starting timing for: iteration 5
Winrate is 68813 and 31187 out of 100000.
The time was 00:00:00.6906353 for: iteration 5
Starting timing for: iteration 6
Winrate is 71332 and 28668 out of 100000.
The time was 00:00:00.7256819 for: iteration 6
Starting timing for: iteration 7
Winrate is 72206 and 27794 out of 100000.
The time was 00:00:00.7381460 for: iteration 7
Starting timing for: iteration 8
Winrate is 71816 and 28184 out of 100000.
The time was 00:00:00.7187728 for: iteration 8
Starting timing for: iteration 9
Winrate is 72756 and 27244 out of 100000.
The time was 00:00:00.7489090 for: iteration 9
Starting timing for: iteration 10
Winrate is 73219 and 26781 out of 100000.
The time was 00:00:00.7602813 for: iteration 10
Starting timing for: iteration 11
Winrate is 72855 and 27145 out of 100000.
The time was 00:00:00.7726654 for: iteration 11
Starting timing for: iteration 12
Winrate is 72996 and 27004 out of 100000.
The time was 00:00:00.7588148 for: iteration 12
Starting timing for: iteration 13
Winrate is 72459 and 27541 out of 100000.
The time was 00:00:00.7528051 for: iteration 13
Starting timing for: iteration 14
Winrate is 72528 and 27472 out of 100000.
The time was 00:00:00.7562398 for: iteration 14
```

On, the last run it got to 74%.

9:45am. Ok, this is good enough. It just shows that I should not be using Monte Carlo to do RL. I definitely need Zap TD to take place of the critic. Well either it or Rudder.

The next thing I should implement will definitely be PG. Let me get to work on it."